fixes for `fix-toml`
